Resultant

+++ lattice/src/config.rs
@@ -1,7 +1,7 @@
 
 
 
-pub const q: i32 = 12289; // Modulus for the polynomial coefficients
+pub const q: i64 = 12289; // Modulus for the polynomial coefficients
 pub const ndim: usize = 8; // Degree of the polynomial
 pub const sigmamin: f64 = 3.19; // Minimum standard deviation of the Gaussian distribution
 pub const sigmamax: f64 = 3.19; // Maximum standard deviation of the Gaussian distribution
